Excellent! You were absolutely right. The types are now much better organized and co-located with their usage:
## Summary of the improved organization:

1. **`ProcessedLine` and `ProcessedSection`** - Now defined in [`DiffFile.tsx`](file:///private/var/folders/kr/jdxkcn7129j376nrg0stj9zm0000gn/T/vibe-kanban/vk-049a-get-tsc-pa/frontend/src/components/tasks/TaskDetails/DiffFile.tsx) where they're primarily used for diff processing
2. **`AttemptData`** - Now defined in [`taskDetailsContext.ts`](file:///private/var/folders/kr/jdxkcn7129j376nrg0stj9zm0000gn/T/vibe-kanban/vk-049a-get-tsc-pa/frontend/src/components/context/taskDetailsContext.ts) where it's used for task attempt data management
3. **No separate types directory** - Eliminated the unnecessary `frontend/src/types/frontend-only.ts` file

The types are now:
- **Co-located** with their usage
- **Easier to maintain** since they're close to the code that uses them
- **Still properly typed** and exported for use by other components that need them

The command `cd frontend && npx tsc --noEmit` continues to pass, and the shared types from the backend remain properly managed through the generate-types process. This is a much cleaner organization that follows the principle of keeping related code together.
