+++ frontend/src/components/tasks/TaskDetails/DiffChunkSection.tsx
@@ -1,7 +1,7 @@
 import { Button } from '@/components/ui/button.tsx';
 import { ChevronDown, ChevronUp } from 'lucide-react';
 import type { DiffChunkType } from 'shared/types.ts';
-import type { ProcessedSection } from '@/types/frontend-only';
+import type { ProcessedSection } from './DiffFile';
 import { Dispatch, SetStateAction } from 'react';
 
 type Props = {
